fix: Namespaced routeOption
