+++ spec/graphql/types/tfl/entities/journey_planner/route_option_type_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe Types::RouteOptionType do
+describe Types::Tfl::Entities::JourneyPlanner::RouteOptionType do
   subject { described_class }
 
   it { is_expected.to have_field(:directions).of_type('[String]') }
